Add plugin that interfaces with website
Allow archiving multiple times to the same park entry
Admin dashboard allows for choosing from any of the saves that were archived during a server's uptime
Auto-archive option, archives when all players disconnect

+++ web.js
@@ -8,6 +8,7 @@ const crypto = require('crypto');
 const path = require('path');
 const fs = require('fs');
 const fsp = fs.promises;
+const net = require('net');
 const DB = require('./db');
 const GameServer = require('./gameserver');
 const FileMan = require('./fileMan');
@@ -44,6 +45,7 @@ class Web {
         const db = this._db = new DB(options.db);
         const port = process.env.PORT || options.port || 8080;
         const privateport = process.env.PRIVATEPORT || options.privateport || 8081;
+        const pluginport = process.env.PLUGINPORT || options.pluginport || 35712;
         this._screenshotter = process.env.SCREENSHOTTER || options.screenshotter || 'screenshotterhost';
         this._archive = options.archivedir || 'storage/archive';
         this._servers = [];
@@ -295,8 +297,10 @@ class Web {
                 });
         });
 
-        privateapp.get('/park/:park', (req, res) => {
+        privateapp.get('/park/:park', async (req, res) => {
             let park = db.GetPark(parseInt(req.params.park));
+            let files = await fsp.readdir(path.join(this._archive, park.dir), { withFileTypes: true });
+            files = files.filter(f => (f.isFile() && f.name.toLowerCase().endsWith('.park'))).map(f => f.name).reverse();
             if (park) {
                 res.render('admin/template',
                     {
@@ -304,7 +308,8 @@ class Web {
                             view: 'park',
                             title: `Park #${req.params.park} - ${park.name} - ${park.groupname} ${park.gamemode} - ${(new Date(park.date)).toLocaleDateString()}`
                         },
-                        park
+                        park,
+                        files
                     },
                     function (err, html) {
                         if (!err) {
@@ -369,7 +374,7 @@ class Web {
             });
         });
 
-        app.get('/api/park/:park', (req, res) => {
+        app.get('/api/park/:park/?', (req, res) => {
             res.send(this.InjectStatus(this._db.GetPark(parseInt(req.params.park) || 0), 'good'));
         });
 
@@ -403,18 +408,21 @@ class Web {
                     const server = servers[serverindx];
                     let filename = false;
                     let dirname = `${datestring}_${server._name}`;
+                    if (server._id) {
+                        dirname = (await db.GetPark(server._id)).dir;
+                    }
                     let archivepath = path.join(this._archive, dirname);
                     try {
-                        if (!(await exists(archivepath))
-                            && !(await fsp.mkdir(archivepath))) {
+                        if (server._id || (!(await exists(archivepath))
+                            && !(await fsp.mkdir(archivepath)))) {
                             for (let i = 0; i < 2; i++) {
                                 filename = await server.SavePark(archivepath);
                                 if (filename) {
                                     break;
                                 }
                             }
-                            if(!filename){
-                                await fsp.rm(archivepath, {recursive: true, force: true});
+                            if (!filename && !server._id) {
+                                await fsp.rm(archivepath, { recursive: true, force: true });
                             }
                         }
                     }
@@ -426,8 +434,12 @@ class Web {
                         result.status = 'bad';
                         status = 500;
                     }
+                    else if (server._id) {
+                        db.ChangeFileName(server._id, filename);
+                        db.RemoveImages(server._id);
+                    }
                     else {
-                        this._db.AddPark({
+                        let result = this._db.AddPark({
                             name: server._name,
                             group: server._group,
                             gamemode: server._mode,
@@ -435,10 +447,13 @@ class Web {
                             dir: dirname,
                             filename
                         });
+                        server._id = result.lastInsertRowid;
                     }
                 }
             }
-            res.status(status).send(result);
+            if (res) {
+                res.status(status).send(result);
+            }
         };
 
         app.get('/api/group/save/:group', async (req, res) => {
@@ -543,6 +558,49 @@ class Web {
             res.status(status).send(result);
         });
 
+        privateapp.post('/api/park/:park/save', async (req, res) => {
+            let park = db.GetPark(parseInt(req.params.park));
+            let message = req.body;
+            let result = {
+                status: 'bad'
+            };
+            let status = 400;
+
+            if (park) {
+                if (req.body.action === 'select' && 'file' in message) {
+                    db.ChangeFileName(park.id, message.file);
+                    db.RemoveImages(park.id);
+                    result.status = 'ok';
+                    status = 200;
+                }
+                else if (req.body.action === 'rm' && 'file' in message && message.file !== park.filename) {
+                    try {
+                        await fsp.unlink(path.join(this._archive, park.dir, message.file));
+                        result.status = 'ok';
+                        status = 200;
+                    }
+                    catch (ex) {
+                        console.error(`Failed to remove save: ${ex}`);
+                    }
+                }
+                else if (req.body.action === 'rm-all') {
+                    let files = await fsp.readdir(path.join(this._archive, park.dir), { withFileTypes: true });
+                    files = files.filter(f => f.isFile() && f.name.toLowerCase().endsWith('.park') && f.name !== park.filename);
+                    let promises = [];
+                    files.forEach(f => promises.push(fsp.unlink(path.join(this._archive, park.dir, f.name))));
+                    try {
+                        await Promise.all(promises);
+                        result.status = 'ok';
+                        status = 200;
+                    }
+                    catch (ex) {
+                        console.error(`Failed to remove all non-selected saves: ${ex}`);
+                    }
+                }
+            }
+            res.status(status).send(result);
+        });
+
         privateapp.post('/api/group/:group/send', async (req, res) => {
             const message = req.body.message;
             const servers = this._servers.filter(server => server._group == req.params.group);
@@ -637,6 +695,18 @@ class Web {
                     let filenamenew = park.name.substring(0, Math.min(fextsep, 25)) + fext;
                     let fullpathnew = path.join(this._archive, parkentry.dir, filenamenew);
                     await fsp.unlink(fullpathold);
+
+                    let files = await fsp.readdir(path.join(this._archive, parkentry.dir), { withFileTypes: true });
+                    files = files.filter(f => f.isFile() && f.name.toLowerCase().endsWith('.park'));
+                    let promises = [];
+                    files.forEach(f => promises.push(fsp.unlink(path.join(this._archive, parkentry.dir, f.name))));
+                    try {
+                        await Promise.all(promises);
+                    }
+                    catch (ex) {
+                        console.error(`Failed to remove all saves while uploading: ${ex}`);
+                    }
+
                     await park.mv(fullpathnew);
 
                     db.ChangeFileName(parkentry.id, filenamenew);
@@ -746,8 +816,39 @@ class Web {
 
         privateapp.use('/', app);
 
+        let pluginserver = new net.Server();
+        pluginserver.on('connection', async sock => {
+            let addr = sock.remoteAddress;
+            let server = null;
+            for (let s in this._servers) {
+                let srv = this._servers[s];
+                if (addr.includes(await srv.GetIP())) {
+                    server = srv;
+                    break;
+                }
+            }
+            if (server) {
+                sock.on('data', async data => {
+                    let payload = JSON.parse(data);
+                    if (payload.type === 'newpark') {
+                        server._id = null;
+                        sock.write('done');
+                    }
+                    if (payload.type === 'archive') {
+                        await saveServers(null, null, [server], false);
+                        sock.write('done');
+                    }
+                });
+            }
+            else {
+                console.error(`Got a plugin connection from unknown IP ${addr}`);
+                sock.close();
+            }
+        });
+
         this._webserver = app.listen(port, () => console.log(`ffa-tycoon running on port ${port}`));
         this._privwebserver = privateapp.listen(privateport, () => console.log(`private backend running on port ${privateport}`));
+        this._pluginsocket = pluginserver.listen(pluginport, () => console.log(`plugin server listening on port ${pluginport}`));
     }
 
     InjectStatus = (obj, status) => {
@@ -765,6 +866,7 @@ class Web {
     close() {
         this._webserver.close();
         this._privwebserver.close();
+        this._pluginsocket.close();
     }
 }
